+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- update method to deposit resource
 
 ## [1.3.0] - 2025-11-06
 ### Added